merge patch from Rorzik, for the problem reported in Debian bug #444638: "xsel clears buffer instead of outputting when backgrounded".
This commit applies Rorzik's logic only for the case where both stdin and
stdout are not ttys, eg. when it is being run from a script. However, the
previous logic remains for cases where one of stdin or stdout are not a
tty, eg. "xsel > file.txt" or "echo text | xsel".


git-svn-id: http://svn.kfish.org/xsel/trunk@195 9c49b5d1-7df3-0310-bce2-b7278e68f44c