Fix overflow of supported_targets array
Patch by Hans de Goede:

I'm one of the developers of spice (a remote virtual machine viewing protocol ala vnc / rdp).
One of the features of spice is copy / paste between the guest and the client, while testing
copy and paste with xsel, the spice-client crashed with a BadAtom error. This is caused
by the spice-client getting the TARGETS property, and then for debugging purposes doing a
XGetAtomName on all reported target Atoms. xsel however claims to send 9 targets, but
only send 8 Atoms + 1 piece of random data from memory.