Send correct SelectionNotify event when starting INCR transfer
The target should be copied from that given in SelectionRequest, not set
to INCR.  This fixes a bug where it was impossible to paste > 4000
characters of text from xsel to the Chromium browser.  Note that some
programs, including Chromium, validate the 'target' that is passed back;
while other programs, such as those which use the GTK clipboard API,
ignore it.  But based the ICCCM, it appears that xsel is incorrect, not
chromium:

    "The owner should set the specified selection, target, time, and
    propety arguments to the values received in the SelectionRequest event."